UART added

+++ utils/custom_types.h
@@ -11,8 +11,8 @@ typedef unsigned char u8_t;
 typedef signed char s8_t;
 typedef unsigned short int u16_t;
 typedef signed short int s16_t;
-typedef unsigned  int u32_t;
-typedef signed  int s32_t;
+typedef unsigned long  int u32_t;
+typedef signed long int s32_t;
 typedef enum
 {
 	false=0,
